// web/src/generated/models/matterSearchResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Priv-View Core API
 * Phase-one Core API for authentication, hierarchical tenants, clients, matters, and matter metadata definitions.
 * OpenAPI spec version: 0.1.0
 */
import type { MatterSearchHit } from './matterSearchHit';
import type { MatterSearchResponseFacets } from './matterSearchResponseFacets';

export interface MatterSearchResponse {
  /** @minimum 0 */
  total: number;
  /** @minimum 0 */
  took_ms: number;
  timed_out: boolean;
  hits: MatterSearchHit[];
  facets: MatterSearchResponseFacets;
}